feat(backend): supports notify watcher for library
