+++ crates/backend/Cargo.toml
@@ -16,3 +16,4 @@ serde = { version = "1.0.114", features = ["derive"] }
 cargo_metadata = "0.10.0"
 wasm-bindgen-cli-support = "0.2.63"
 toml = "0.5.6"
+notify = "4.0.15"
