+++ crates/backend/tests/dir.rs
@@ -6,7 +6,6 @@ fn test_build() {
     let root = PathBuf::from(".");
     let pkg = Crate::new(root).unwrap();
 
-    // println!("{:?}", env::var("OUT_DIR").unwrap());
     pkg.build().unwrap();
     pkg.bindgen().unwrap();
 }
